LayoutTests/fast/profiler/anonymous-functions-with-display-names-expected.txt

This page's JavaScript calls anonymous functions with display names. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. You should see functions named "[x] iterations function", and no anonymous functions.

Profile title: Anonymous functions with display names
Thread_1 (no file) (line 0:0)
   startTest anonymous-functions-with-display-names.html (line 22:19)
      anonymousFunctionGenerator anonymous-functions-with-display-names.html (line 10:36)
      0 iterations function anonymous-functions-with-display-names.html (line 12:59)
      1 iterations function anonymous-functions-with-display-names.html (line 12:59)
      2 iterations function anonymous-functions-with-display-names.html (line 12:59)
      20 iterations function anonymous-functions-with-display-names.html (line 12:59)
      1000 iterations function anonymous-functions-with-display-names.html (line 12:59)
      endTest profiler-test-JS-resources.js (line 1:17)
